# api/dist/openapi.yaml
openapi: 3.1.0
info:
  title: Nerve Wiki API
  version: v0
  description: The HTTP API of Nerve Wiki. The web app is one client of it; every account can use every endpoint the same way. Every operation lists the problem codes it can answer in x-problem-codes; the top-level x-problem-codes can come from every operation, and an operation that needs a bearer token can also answer unauthorized.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
...
paths:
...
  /api/v0/notebooks/{notebook_id}/assets:
    parameters:
      - $ref: '#/components/parameters/NotebookID'
    post:
      operationId: uploadAsset
      tags:
        - asset
      summary: Upload an attachment
      description: 'Uploads a file as an attachment of the notebook, under a page or at its root, last among its siblings; its notebook''s editors and admins can, a reader cannot (forbidden). The body is multipart/form-data whose parts are parent_id, name and file, in that order, the first two optional: an unknown part, one twice, one out of order, a part after the file, a parent_id that is no id, a name of more than 1 KiB or not UTF-8, more than 4 KiB read before the file''s bytes, no file, and a body that goes on after the form past the limit are bad_request. Each part is read as sent: a Content-Transfer-Encoding is not decoded. Without a name, or with one empty or of blanks, the file part''s file name is taken. Before the file is read, a notebook that does not exist, is deleted, or that the caller has no role in is notebook.not_found; the name follows a page''s title''s rules and does not end with ".md", in any case, and the parent is a page of the notebook (validation_failed); a name a sibling has, compared by its key, is page.title_taken; and a server whose storage keeps no more room is storage_full: a client answered before it sent its file may see the connection reset, and checks these first. A file larger than the instance''s asset_max_bytes is payload_too_large; one that does not arrive at the instance''s lowest rate is cut off, as is one still arriving when the server shuts down. A file the storage runs out of room for is storage_full. The checks are made again as the attachment is created. The server tells the file''s type from its name''s extension and its first bytes, and reads an image''s size.'
      security:
        - bearer: []
      x-raw: true
      x-problem-codes:
        - notebook.not_found
        - forbidden
        - validation_failed
        - page.title_taken
        - storage_full
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required:
                - file
              properties:
                parent_id:
                  description: The page to put the attachment under; none puts it at the notebook's root.
                  type: string
                  format: uuid
                name:
                  description: The attachment's name; none, or one empty or of blanks, takes the file part's file name.
                  type: string
                file:
                  description: The file, the last part.
                  type: string
                  format: binary
      responses:
        '201':
          description: The attachment, uploaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        default:
          $ref: '#/components/responses/Problem'
    get:
      operationId: listAssets
      tags:
        - asset
      summary: List the attachments under a page
      description: The attachments under a page of the notebook, or at its root without parent_id, by name, compared by its key, then id, a page at a time, each with its addresses signed. Any role in the notebook can list them. A cursor the list cannot read is bad_request, before anything else; a notebook that does not exist, is deleted, or that the caller has no role in is notebook.not_found; a parent_id that is no page of the notebook, or a deleted one, is page.not_found; a limit outside 1–100 is validation_failed.
      security:
        - bearer: []
      x-problem-codes:
        - bad_request
        - notebook.not_found
        - page.not_found
        - validation_failed
      parameters:
        - name: parent_id
          in: query
          description: The page whose attachments to list; absent for the notebook's root.
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: A page of the attachments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPage'
        default:
          $ref: '#/components/responses/Problem'
  /api/v0/assets/{node_id}:
    parameters:
      - $ref: '#/components/parameters/NodeID'
    get:
      operationId: getAsset
      tags:
        - asset
      summary: Read an attachment
      description: The attachment, its addresses signed. Any role in its notebook can read it. A node that does not exist, is deleted, is no attachment, or whose notebook the caller has no role in is asset.not_found.
      security:
        - bearer: []
      x-problem-codes:
        - asset.not_found
      responses:
        '200':
          description: The attachment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: An access token (JWT) from register, login or refresh, or a personal access token (nwk_pat_) from createApiToken.
  schemas:
...
    FieldError:
      description: One invalid field of a request. Clients show text looked up by `code`; `message` is an English explanation for developers. Must match httpserver.FieldError and the field codes of internal/shared.
      type: object
      additionalProperties: false
      required:
        - field
        - code
        - message
      properties:
        field:
          description: The field's path in the request, e.g. title or properties.status. A path longer than 256 bytes is cut short and ends with `…`.
          type: string
        code:
          description: What is wrong with the field.
          type: string
          enum:
            - required
            - invalid_format
            - too_short
            - too_long
            - out_of_range
            - not_allowed
            - duplicate
            - common_password
        message:
          type: string
    Problem:
      description: RFC 9457 problem details. `title` is the HTTP status phrase, `detail` explains this occurrence, and clients branch on `code`. Must match httpserver.Problem; the platform's contract test checks it.
      type: object
      additionalProperties: false
      required:
        - status
        - code
        - title
      properties:
        status:
          description: HTTP status code.
          type: integer
        code:
          description: Stable error code. Platform codes have no prefix (bad_request, unauthorized, forbidden, not_found, payload_too_large, validation_failed, rate_limited, server_busy, internal_error, not_ready, storage_full); module codes are prefixed with the module that refuses, e.g. page.locked. Each operation lists the codes it can answer in x-problem-codes.
          type: string
        title:
          description: HTTP status phrase, e.g. "Not Found".
          type: string
        detail:
          description: What went wrong in this occurrence.
          type: string
        errors:
          description: The invalid fields of the request.
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        lock:
          description: 'The edit lock page.locked names: the page locked and who holds it (the caller themself when they hold it elsewhere). No other code carries it.'
          type: object
          additionalProperties: false
          required:
            - page_id
            - user_id
            - display_name
          properties:
            page_id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            display_name:
              type: string
        locks:
          description: 'The edit locks linking.pages_locked names, one a page: each page whose links a rename or move would write again and who edits it (the caller themself when they edit it). No other code carries it.'
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - page_id
              - user_id
              - display_name
            properties:
              page_id:
                type: string
                format: uuid
              user_id:
                type: string
                format: uuid
              display_name:
                type: string
        ended_by:
          description: 'Who ended the caller''s edit session: page.edit_session_unlocked names the notebook''s admin who unlocked it. No other code carries it.'
          type: object
          additionalProperties: false
          required:
            - user_id
            - display_name
          properties:
            user_id:
              type: string
              format: uuid
            display_name:
              type: string
...
    NextCursor:
      description: The cursor of the next page; null on the last page.
      type:
        - string
        - 'null'
...
    Asset:
      description: 'An attachment: its node in the notebook''s tree, its file, and the addresses of its content, signed for anyone who has them.'
      type: object
      additionalProperties: false
      required:
        - id
        - notebook_id
        - parent_id
        - name
        - link
        - mime
        - byte_size
        - sha256
        - width
        - height
        - created_by
        - created_at
        - content_url
        - download_url
        - expires_at
      properties:
        id:
          description: The attachment's node.
          type: string
          format: uuid
        notebook_id:
          type: string
          format: uuid
        parent_id:
          description: The page it is under; null at the notebook's root.
          type:
            - string
            - 'null'
          format: uuid
        name:
          description: The attachment's file name.
          type: string
        link:
          description: 'How a wikilink, ![[link]] to embed it, is written to lead to the attachment alone from anywhere in the notebook, as the link targets write it: its name, or its path from the root where another attachment has its name''s title key. It is as of the answer: another attachment given that title key later may make the name lead elsewhere. Null for an attachment without an extension, which no link leads to.'
          type:
            - string
            - 'null'
        mime:
          description: 'The type the server told from the name''s extension and the file''s first bytes, which it serves the content as: an image, an audio, a video or a PDF it shows; application/octet-stream for any other file, which is downloaded.'
          type: string
        byte_size:
          description: The file's size in bytes.
          type: integer
          format: int64
        sha256:
          description: The file's SHA-256, in lower-case hexadecimal.
          type: string
        width:
          description: An image's width in pixels (PNG, JPEG, GIF), when the server read it; null otherwise.
          type:
            - integer
            - 'null'
        height:
          description: An image's height in pixels, as width.
          type:
            - integer
            - 'null'
        created_by:
          description: Who uploaded it.
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        content_url:
          description: 'The address of the content, on this server, which a browser opens without a token until expires_at: an image, an audio, a video or a PDF is shown, any other file downloaded.'
          type: string
        download_url:
          description: The address of the content to download, whatever its type, until expires_at.
          type: string
        expires_at:
          description: When the addresses stop working, one to two hours after they were signed; reading the attachment again signs new ones.
          type: string
          format: date-time
    AssetPage:
      type: object
      additionalProperties: false
      required:
        - data
        - next_cursor
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        next_cursor:
          $ref: '#/components/schemas/NextCursor'
  responses:
    Problem:
      description: Error (RFC 9457 problem details).
      headers:
        Retry-After:
          description: Whole seconds to wait before trying again, rounded up; sent with rate_limited, server_busy and not_ready.
          schema:
            type: integer
            minimum: 1
        WWW-Authenticate:
          description: 'Sent with every 401 (RFC 9110 15.5.2): Bearer error="invalid_token" (RFC 6750 3) when the bearer token sent is refused before the operation runs, being invalid or expired; plain Bearer for every other 401, including a credential that the operation finds revoked while it runs.'
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  parameters:
...
    NotebookID:
      name: notebook_id
      in: path
      required: true
      description: The notebook's id.
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      description: The page size, 1–100; 50 when absent. Outside that range the answer is 422 validation_failed on limit.
      schema:
        type: integer
    Cursor:
      name: cursor
      in: query
      description: The next_cursor of the page before; absent for the first page. A cursor that does not decode, has an unknown version or a payload of another shape than this list's, or is not spelled as the server writes it is 400 bad_request on cursor.
      schema:
        type: string
...
    NodeID:
      name: node_id
      in: path
      required: true
      description: The id of a node of a notebook's tree.
      schema:
        type: string
        format: uuid
...