# server/internal/modules/asset/adapter/http/gen/server.gen.go
//go:build go1.22

// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"net/http"
	"time"
	"uuid"

	"github.com/oapi-codegen/nullable"
	"github.com/oapi-codegen/runtime"
	externalRef0 "github.com/open-nerve/NerveWiki/server/internal/platform/httpserver/apigen"
)

// Asset An attachment: its node in the notebook's tree, its file, and the addresses of its content, signed for anyone who has them.
type Asset struct {
	// ByteSize The file's size in bytes.
	ByteSize int64 `json:"byte_size"`

	// ContentURL The address of the content, on this server, which a browser opens without a token until expires_at: an image, an audio, a video or a PDF is shown, any other file downloaded.
	ContentURL string    `json:"content_url"`
	CreatedAt  time.Time `json:"created_at"`

	// CreatedBy Who uploaded it.
	CreatedBy uuid.UUID `json:"created_by"`

	// DownloadURL The address of the content to download, whatever its type, until expires_at.
	DownloadURL string `json:"download_url"`

	// ExpiresAt When the addresses stop working, one to two hours after they were signed; reading the attachment again signs new ones.
	ExpiresAt time.Time `json:"expires_at"`

	// Height An image's height in pixels, as width.
	Height nullable.Nullable[int] `json:"height"`

	// ID The attachment's node.
	ID uuid.UUID `json:"id"`

	// Link How a wikilink, ![[link]] to embed it, is written to lead to the attachment alone from anywhere in the notebook, as the link targets write it: its name, or its path from the root where another attachment has its name's title key. It is as of the answer: another attachment given that title key later may make the name lead elsewhere. Null for an attachment without an extension, which no link leads to.
	Link nullable.Nullable[string] `json:"link"`

	// Mime The type the server told from the name's extension and the file's first bytes, which it serves the content as: an image, an audio, a video or a PDF it shows; application/octet-stream for any other file, which is downloaded.
	Mime string `json:"mime"`

	// Name The attachment's file name.
	Name       string    `json:"name"`
	NotebookID uuid.UUID `json:"notebook_id"`

	// ParentID The page it is under; null at the notebook's root.
	ParentID nullable.Nullable[uuid.UUID] `json:"parent_id"`

	// Sha256 The file's SHA-256, in lower-case hexadecimal.
	Sha256 string `json:"sha256"`

	// Width An image's width in pixels (PNG, JPEG, GIF), when the server read it; null otherwise.
	Width nullable.Nullable[int] `json:"width"`
}

// AssetPage defines model for AssetPage.
type AssetPage struct {
	Data []Asset `json:"data"`

	// NextCursor The cursor of the next page; null on the last page.
	NextCursor nullable.Nullable[externalRef0.NextCursor] `json:"next_cursor"`
}

// NodeID defines model for NodeID.
type NodeID = uuid.UUID

// NotebookID defines model for NotebookID.
type NotebookID = uuid.UUID

// Problem RFC 9457 problem details. `title` is the HTTP status phrase, `detail` explains this occurrence, and clients branch on `code`. Must match httpserver.Problem; the platform's contract test checks it.
type Problem = externalRef0.Problem

// ListAssetsParams defines parameters for ListAssets.
type ListAssetsParams struct {
	// ParentID The page whose attachments to list; absent for the notebook's root.
	ParentID *uuid.UUID `form:"parent_id,omitempty" json:"parent_id,omitempty"`

	// Limit The page size, 1–100; 50 when absent. Outside that range the answer is 422 validation_failed on limit.
	Limit *externalRef0.Limit `form:"limit,omitempty" json:"limit,omitempty"`

	// Cursor The next_cursor of the page before; absent for the first page. A cursor that does not decode, has an unknown version or a payload of another shape than this list's, or is not spelled as the server writes it is 400 bad_request on cursor.
	Cursor *externalRef0.Cursor `form:"cursor,omitempty" json:"cursor,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetAsset Read an attachment
	// (GET /api/v0/assets/{node_id})
	GetAsset(w http.ResponseWriter, r *http.Request, nodeID NodeID)
	// ListAssets List the attachments under a page
	// (GET /api/v0/notebooks/{notebook_id}/assets)
	ListAssets(w http.ResponseWriter, r *http.Request, notebookID NotebookID, params ListAssetsParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetAsset operation middleware
func (siw *ServerInterfaceWrapper) GetAsset(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "node_id" -------------
	var nodeID NodeID

	err = runtime.BindStyledParameterWithOptions("simple", "node_id", r.PathValue("node_id"), &nodeID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "node_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAsset(w, r, nodeID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListAssets operation middleware
func (siw *ServerInterfaceWrapper) ListAssets(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "notebook_id" -------------
	var notebookID NotebookID

	err = runtime.BindStyledParameterWithOptions("simple", "notebook_id", r.PathValue("notebook_id"), &notebookID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "notebook_id", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params ListAssetsParams

	// ------------- Optional query parameter "parent_id" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "parent_id", r.URL.Query(), &params.ParentID, runtime.BindQueryParameterOptions{Type: "string", Format: "uuid"})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "parent_id"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "parent_id", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "cursor" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "cursor", r.URL.Query(), &params.Cursor, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "cursor"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "cursor", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListAssets(w, r, notebookID, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v0/notebooks/{notebook_id}/assets", wrapper.ListAssets)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/v0/assets/{node_id}", wrapper.GetAsset)

	return m
}

type ProblemResponseHeaders struct {
	RetryAfter      *int
	WWWAuthenticate *string
}
type ProblemApplicationProblemPlusJSONResponse struct {
	Body externalRef0.Problem

	Headers ProblemResponseHeaders
}

type GetAssetRequestObject struct {
	NodeID NodeID `json:"node_id"`
}

type GetAssetResponseObject interface {
	VisitGetAssetResponse(w http.ResponseWriter) error
}

type GetAsset200JSONResponse Asset

func (response GetAsset200JSONResponse) VisitGetAssetResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetAssetdefaultApplicationProblemPlusJSONResponse struct {
	Body       externalRef0.Problem
	Headers    ProblemResponseHeaders
	StatusCode int
}

func (response GetAssetdefaultApplicationProblemPlusJSONResponse) VisitGetAssetResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	if response.Headers.RetryAfter != nil {
		w.Header().Set("Retry-After", fmt.Sprint(*response.Headers.RetryAfter))
	}
	if response.Headers.WWWAuthenticate != nil {
		w.Header().Set("WWW-Authenticate", fmt.Sprint(*response.Headers.WWWAuthenticate))
	}
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type ListAssetsRequestObject struct {
	NotebookID NotebookID `json:"notebook_id"`
	Params     ListAssetsParams
}

type ListAssetsResponseObject interface {
	VisitListAssetsResponse(w http.ResponseWriter) error
}

type ListAssets200JSONResponse AssetPage

func (response ListAssets200JSONResponse) VisitListAssetsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type ListAssetsdefaultApplicationProblemPlusJSONResponse struct {
	Body       externalRef0.Problem
	Headers    ProblemResponseHeaders
	StatusCode int
}

func (response ListAssetsdefaultApplicationProblemPlusJSONResponse) VisitListAssetsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	if response.Headers.RetryAfter != nil {
		w.Header().Set("Retry-After", fmt.Sprint(*response.Headers.RetryAfter))
	}
	if response.Headers.WWWAuthenticate != nil {
		w.Header().Set("WWW-Authenticate", fmt.Sprint(*response.Headers.WWWAuthenticate))
	}
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetAsset Read an attachment
	// (GET /api/v0/assets/{node_id})
	GetAsset(ctx context.Context, request GetAssetRequestObject) (GetAssetResponseObject, error)
	// ListAssets List the attachments under a page
	// (GET /api/v0/notebooks/{notebook_id}/assets)
	ListAssets(ctx context.Context, request ListAssetsRequestObject) (ListAssetsResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetAsset operation middleware
func (sh *strictHandler) GetAsset(w http.ResponseWriter, r *http.Request, nodeID NodeID) {
	var request GetAssetRequestObject

	request.NodeID = nodeID

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetAsset(ctx, request.(GetAssetRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetAsset")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetAssetResponseObject); ok {
		if err := validResponse.VisitGetAssetResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ListAssets operation middleware
func (sh *strictHandler) ListAssets(w http.ResponseWriter, r *http.Request, notebookID NotebookID, params ListAssetsParams) {
	var request ListAssetsRequestObject

	request.NotebookID = notebookID
	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListAssets(ctx, request.(ListAssetsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListAssets")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListAssetsResponseObject); ok {
		if err := validResponse.VisitListAssetsResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}